Add navbar, header, graphics, quote
I created the base HTML for the page and then styled it according to
the color and font styles I was given. I should have done a commit
earlier (getting used to commits). I'm at the point where I just have
the CTA section and footer to style. Then I'll go back and make little
optimizations to make sure everything is in order.